API/requirements.txt
fastapi
uvicorn
numpy
pillow
python-multipart
gtts
playsound
torch
torchvision
matplotlib
